add order pages

+++ db/migrate/20141029144823_create_payments.rb
@@ -1,7 +1,7 @@
 class CreatePayments < ActiveRecord::Migration
   def change
     create_table :payments do |t|
-      t.integer    :number
+      t.decimal    :number
       t.date       :expiration_date
       t.integer    :card_code
